declare appInsights types in appInsights module
